The problem was that when you tried to make a change and send the color change to the alienfx library an exception was thrown saying the size should be nine but it was returned as -9.  The change to fix this was to comment out the implementation of wairForReady().  This is just not working and never returns. Commenting it out lets the system function.
The second problem was that RED worked. GREEN worked but only when selecting BLUE.
BLUE did not work at all.

The fix was to do a simple swap of the 'g' and 'b' variables.  Now the GREEN works
when selecting green.

The third change was that though the system was working well for RED and GREEN it
did not work for BLUE.  RED and GREEN are working by the first 4 bits of RG being
used for RED and the last 4 bits being used for green.  ((g >> 4) But blue is now
being controled by only the first 4 bits of the B variable.

The fix was to use
RG
BR
GB
I needed to add a new variable GB and change B to BR to show what they are doing.
Then I was able to use the second half of each of these variables (>> 4) for the
appropriate colors.  Then this new variable was put into the last parameter of a
call to the LEDController.  Now all colors work and are selectable from the GUI.